Move abridgeDescription() to /common/, <hr> style to SCSS module
