+++ src/components/DownloadsPreview.module.scss
@@ -63,6 +63,9 @@ a, Link {
         text-align: justify;
         color: white;
     }
+    hr {
+        margin: 8px 0;
+    }
 }
 
 .description {
